Make authorization optional for some Albums Apis
GET api/albums
If authorization token is NOT provided then all the albums that are NOT private are returned. If such token is proived, then the result includes all the private album of the authenticated user
GET api/albums/{id}
Gets the album by id if it is not private, or if it is private gets it if it is owned by the authenitcated user.